+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2022-12-21
+
+### Added
+
+- `TomlExporter` component for exporting TOML files from Unity scenes
+- Documentation for `TomlExporter` component
+- Exporter sample scene
+- Serializer sample scene
+
+### Fixed
+
+- `TomlSerializer` not putting a new line space between tables
+
 ## [0.1.0] - 2022-12-20
 
 ### Added